+++ payload/install-openhtpc-fedora.sh
@@ -2,7 +2,7 @@
 
 set -Eeuo pipefail
 
-readonly OPENHTPC_VERSION="1.2.0-dev8"
+readonly OPENHTPC_VERSION="1.2.0-dev9"
 
 
 readonly SCRIPT_DIR="$(cd -- "$(dirname -- "${BASH_SOURCE[0]}")" && pwd)"